v0.0.7

+++ Models/Inmuebles.cs
@@ -9,8 +9,8 @@ public class Inmuebles
     public string Tipo { get; set; } = "";
     public int Ambiente { get; set; }
     public decimal Precio { get; set; }
-    public float Longitud { get; set; }
-    public float Latitud { get; set; }
+    public double Longitud { get; set; }
+    public double Latitud { get; set; }
     public DateTime Fecha_creacion { get; set; }
     public Boolean Estado { get; set; }
